requirements.txt
numpy==1.26.0
opencv-python==4.8.1.78
# packaging==23.1
Pillow==10.0.1
pytesseract==0.3.10
